Use responders API field
Create a responders field from the users, teams, escalations, and
schedules values in Sensu. The teams value can be used as before and the
other three work the same way. The correct type key is added for you.

Because the old recipients value could contain any of the above, I
haven't tried to map it to the new values so this is a backwards
incompatible change as it won't use it any longer, however the teams
value still works as before.